releases/v2.5.0/requirements.txt
telethon==1.34.0
openai>=1.30.0
python-dotenv==1.0.0
httpx>=0.27.0
streamlit>=1.30.0
psutil>=5.9.0
pypdf>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
